# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "Artifact": {
        "additionalProperties": false,
        "properties": {
          "assignment_id": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 1,
                "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assignment Id"
          },
          "conversation_id": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 1,
                "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversation Id"
          },
          "current_revision": {
            "$ref": "#/components/schemas/Revision"
          },
          "current_revision_id": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]*$",
            "title": "Current Revision Id",
            "type": "string"
          },
          "id": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]*$",
            "title": "Id",
            "type": "string"
          },
          "observed_at": {
            "format": "date-time",
            "title": "Observed At",
            "type": "string"
          },
          "requested_revision": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Revision"
              },
              {
                "type": "null"
              }
            ]
          },
          "workspace_id": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]*$",
            "title": "Workspace Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "workspace_id",
          "current_revision_id",
          "current_revision"
        ],
        "title": "Artifact",
        "type": "object"
      },
...
      "ErrorCode": {
        "enum": [
          "not_found_or_not_authorized",
          "version_conflict",
          "source_changed",
          "source_unavailable",
          "decision_required",
          "decision_stale",
          "budget_exhausted",
          "connection_required",
          "unsupported_operation",
          "action_unresolved",
          "command_conflict",
          "unauthenticated",
          "validation_error",
          "internal_error"
        ],
        "title": "ErrorCode",
        "type": "string"
      },
      "ErrorDetails": {
        "additionalProperties": false,
        "properties": {
          "current_revision_id": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 1,
                "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Revision Id"
          },
          "current_version": {
            "anyOf": [
              {
                "maximum": 2147483647.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Version"
          },
          "fields": {
            "items": {
              "type": "string"
            },
            "maxItems": 50,
            "title": "Fields",
            "type": "array"
          },
          "proposal_id": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 1,
                "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proposal Id"
          }
        },
        "title": "ErrorDetails",
        "type": "object"
      },
      "ErrorEnvelope": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "details": {
            "$ref": "#/components/schemas/ErrorDetails"
          },
          "message": {
            "maxLength": 10000,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          },
          "next_action": {
            "maxLength": 10000,
            "minLength": 1,
            "title": "Next Action",
            "type": "string"
          },
          "request_id": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]*$",
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "request_id",
          "next_action"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "Workspace": {
        "additionalProperties": false,
        "properties": {
          "access_generation": {
            "default": 1,
            "maximum": 2147483647.0,
            "minimum": 1.0,
            "title": "Access Generation",
            "type": "integer"
          },
          "id": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]*$",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "default": "personal",
            "enum": [
              "personal",
              "shared"
            ],
            "title": "Kind",
            "type": "string"
          },
          "name": {
            "maxLength": 10000,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "owner_id": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]*$",
            "title": "Owner Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "owner_id"
        ],
        "title": "Workspace",
        "type": "object"
      },
      "WorkspacePage": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Workspace"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 1,
                "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "WorkspacePage",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Isolated PostgreSQL-backed fixture/scenario service. Not a live provider runtime. Bearer is a private local human credential; no model-callable approval surface.",
    "title": "Workagent local domain baseline",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/v1/workspaces": {
      "get": {
        "operationId": "list_workspaces",
        "parameters": [
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "minLength": 1,
                  "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]*$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "x-schema-version",
            "required": true,
            "schema": {
              "const": "workagent/v1",
              "title": "X-Schema-Version",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-request-id",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]*$",
              "title": "X-Request-Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspacePage"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Missing or invalid local authentication"
          },
          "404": {
            "content": {
              "application/json": {
                "example": {
                  "code": "not_found_or_not_authorized",
                  "details": {
                    "current_revision_id": null,
                    "current_version": null,
                    "fields": [],
                    "proposal_id": null
                  },
                  "message": "Resource unavailable.",
                  "next_action": "Check your workspace and current access.",
                  "request_id": "request-example"
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Absent and inaccessible resources are indistinguishable"
          },
          "409": {
            "content": {
              "application/json": {
                "examples": {
                  "command_conflict": {
                    "value": {
                      "code": "command_conflict",
                      "details": {
                        "current_revision_id": null,
                        "current_version": null,
                        "fields": [],
                        "proposal_id": null
                      },
                      "message": "This command ID has a different payload.",
                      "next_action": "Use the original payload or a new command ID for a different intent.",
                      "request_id": "request-example"
                    }
                  },
                  "stale": {
                    "value": {
                      "code": "version_conflict",
                      "details": {
                        "current_revision_id": "revision-2",
                        "current_version": null,
                        "fields": [],
                        "proposal_id": "proposal-example"
                      },
                      "message": "The current version changed.",
                      "next_action": "Refresh current work and review a new proposal against that base.",
                      "request_id": "request-example"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Domain conflict, unavailable source, decision, budget or unresolved state"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Schema validation or unsupported operation"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Sanitized internal adapter/storage failure; inspect state before retrying a mutation"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Workspaces"
      }
    },
    "/v1/workspaces/{workspace_id}/artifacts/{artifact_id}": {
      "get": {
        "operationId": "get_artifact",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]*$",
              "title": "Workspace Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "artifact_id",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]*$",
              "title": "Artifact Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "revision_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "minLength": 1,
                  "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]*$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Revision Id"
            }
          },
          {
            "in": "header",
            "name": "x-schema-version",
            "required": true,
            "schema": {
              "const": "workagent/v1",
              "title": "X-Schema-Version",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-request-id",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]*$",
              "title": "X-Request-Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "assignment_id": "assignment-example",
                  "conversation_id": null,
                  "current_revision": {
                    "artifact_id": "artifact-example",
                    "author_id": "local-human",
                    "author_kind": "worker",
                    "body": {
                      "blocks": [
                        {
                          "block_id": "next-action",
                          "checked": null,
                          "kind": "checklist",
                          "text": "Record an owner and a next action for the next personal intake case."
                        }
                      ],
                      "title": "Private working plan"
                    },
                    "body_hash": "23a7209c544757d28fb7399c967216e0ff3e078889d00583a8e94551b76b66d2",
                    "created_at": "2026-10-02T00:00:00Z",
                    "id": "revision-1",
                    "parent_revision_id": null,
                    "revision_number": 1,
                    "source_dependencies": [
                      {
                        "external_version": "1",
                        "observed_at": "2026-10-02T00:00:00Z",
                        "source_id": "SG-F2"
                      }
                    ]
                  },
                  "current_revision_id": "revision-1",
                  "id": "artifact-example",
                  "observed_at": "2026-10-02T00:00:00Z",
                  "requested_revision": null,
                  "workspace_id": "workspace-example"
                },
                "schema": {
                  "$ref": "#/components/schemas/Artifact"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Missing or invalid local authentication"
          },
          "404": {
            "content": {
              "application/json": {
                "example": {
                  "code": "not_found_or_not_authorized",
                  "details": {
                    "current_revision_id": null,
                    "current_version": null,
                    "fields": [],
                    "proposal_id": null
                  },
                  "message": "Resource unavailable.",
                  "next_action": "Check your workspace and current access.",
                  "request_id": "request-example"
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Absent and inaccessible resources are indistinguishable"
          },
          "409": {
            "content": {
              "application/json": {
                "examples": {
                  "command_conflict": {
                    "value": {
                      "code": "command_conflict",
                      "details": {
                        "current_revision_id": null,
                        "current_version": null,
                        "fields": [],
                        "proposal_id": null
                      },
                      "message": "This command ID has a different payload.",
                      "next_action": "Use the original payload or a new command ID for a different intent.",
                      "request_id": "request-example"
                    }
                  },
                  "stale": {
                    "value": {
                      "code": "version_conflict",
                      "details": {
                        "current_revision_id": "revision-2",
                        "current_version": null,
                        "fields": [],
                        "proposal_id": "proposal-example"
                      },
                      "message": "The current version changed.",
                      "next_action": "Refresh current work and review a new proposal against that base.",
                      "request_id": "request-example"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Domain conflict, unavailable source, decision, budget or unresolved state"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Schema validation or unsupported operation"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Sanitized internal adapter/storage failure; inspect state before retrying a mutation"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Artifact"
      }
    },
...
  }
}